Initialize basic setup.
Create an index.html file and fill it with boilerplate. Set the src of the script to main.js and set the type as module so that the file can be loaded as a module.

Create a linked-list.js file and
initialize the LinkedList and Node
classes with basic properties.

In main.js, import linked-lists
and create a new instance of it.
This will be used to test the
LinkedList.