Implement set() on linked list
Initialize a set() method
that takes in a `key` and
a `value`. Create a new Node
and set the new Node as
the head if there is no
head.

If the key of
the head is the same as
the key of this new node,
have the value of this
new node, replace the value
of the head. If not, keep
traversing the linked list
and any Node that bears the
same key as this new node
will have its value replaced
by the value of this new
node.

If none of this is the case,
then append this new node
as the last node.

The value overwriting feature
is to prevent nodes with same
keys to be stored in the
linked list (this is not a
collision)
and this feature
is essential to hashmaps.

Test this method in index.js
and log the list to the
console.