Implement hash()
Initialize hash(), a
simple method that
takes in a string and
returns a number that
depend on where the
letters appear in the
string.

Take the modulo of the
hashCode and capacity
because:

"For very long keys, our
hash code will exceed the
maximum integer value allowed
by JavaScript.
Once that happens,
calculations become
inaccurate, and the
chance of collisions
significantly increases."

Taking the modulo ensures
that the output never gets
larger than the bucket's
length.

Import HashMap in main.js
and test this method.