Implement set() in HashMap.
First the name of set()
in the LinkedList class
to put() so prevent
confusion with HashMap's
set(). Update its reference
in main.js

Import the LinkedList class
to hash-map.js.
Initialize set(), a method
that takes in a `key` and
a `value` that sets the key
and its value as a
linked list in buckets.

Hash the key first and check
whether the index returned from
the hashing of the key exists
in the buckets array. If it
does not, create a new linked
list at the index.

Call the LinkedList's put()
method on the buckets array
at this index while passing
the key and value pair.

In main.js, test the
set() method.