Update put()
The node traversal in
put() was faulty.
The first
currentNode = currentNode.nextNode
in the while loop (before this
commit) was traversing
the next node after the head.
If there is a node after the first
node, this line traverses to the
this second node and checks the
condition. The next
currentNode = currentNode.nextNode
(outside the while loop)
sets the current node to null
if this second node happens to
be the last node.
The while loop then throws an
error because it attemps to
access next node of null.

Taking the traversal on the
first line inisde the while
loop fixes this.