Implement fetch()
In the LinkedList class,
initialize a fetch()
method that takes in a
a `key`argument, traverses the
linked list and returns
the value assigned to the
key or returns null if
the key is not found.

This will be used in
the HashMap.

In the main.js, log the
result of calling this
method.