Implement get()
Initialize a method that
takes a `key` as an argument
and returns the value assigned
to the key. Return null
if the key is not found.

This method calls the fetch()
method of LinkedList on the
linkedlist whose index in the
buckets array matches the
index returned from hashing the
key.

In main.js log the result
of calling the get method.