Implement exists() and has()
In the LinkedLists class,
initialize an exists()
method that takes a `key`
as an
argument. Traverse through
the linked list and return
true if key is found or
false if not found.

In the HashMaps class,
initialize a has()
method that takes a `key`
as an argument. It
returns false if the index
which is the hash result
of the key is not found
on the buckets array.
If the index is found,
it calls the exists()
method on the index
in the buckets array and
returns the result.

In main.js, log the results
of calling these two methods.