Implement length() and clear()
Initialize a length() method that
returns the total of
stored keys in the hashmap.

If there are no keys, return 0.

Else use the array reduce method
on the buckets array and call
the LinkedList's size method
on each bucket (list) and
return the total.

Initialize a clear() method
that removes all enteries in
the hashmap by emptying the
buckets array.