Implement keys()
Initialize a keys()
method that returns an
array of all keys in
the HashMap.

Return the result of
calling the array map method
on the bucket array.
For each bucket, call
the LinkedList getKeys()
method on it. Since the
getKeys() returns an array,
and map returns an array,
the final result will be
an a 2D array. Flatten
it with the flat() method.