Implement entries()
Initialize entries()
which returns an array
of all the key-value
pairs in the HashMap.

The entries() method
calls the keys() and
values() methods.
Call the array map
method on the keys array
and for each key, create an
array containing the key and
its corresponding value
from the values array.
Return the result.

The map() method
runs the callback function
on every element in the
array and returns a new
array containing the
results.
This results in a 2D
array in the following
format:

`[[firstKey, firstValue], [secondKey, secondValue]]`