Implement checkThreshold and rehash
Initialize a checkThreshold method.
Calculate the threshold and
return true if the length of
the hashmap us greater than
the threshold.

Initialize a rehash method
Double the capacity and
clear the hashmap.
Call the array map on the
hashmap's entries
(an array containing
arrays of key-value pair)
and for each entry,
deconstruct it into
a key-value pair and
call the set() method
on the key-value pair.

These methods will be
used to handle the
hashmap growth logic
checkThreshold will
check if the hashmap
has reached its threshold
and if so, the capacity
of the hashmap will
be doubled and all
entries in the hashmap
will be rehashed
(index for each key
is recalculated
based on the new
capacity.)